backend/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [ "setuptools>=80", "setuptools-scm[toml]>=6.2" ]

[project]
name = "forecast-in-a-box"
description = "Weather Forecast in a Box"
readme = "README.md"
license = "Apache-2.0"
license-files = [ "LICENSE" ]
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]
requires-python = ">=3.11"
# [tool.setuptools]
# packages = ["forecastbox"]
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]

# just a minimal set of dependencies to run the servers. Individual tasks install their runtime dependencies on their own
dependencies = [
  "aiosqlite",
  "anemoi-inference==0.8.3",
  "anemoi-plugins-ecmwf-inference[opendata,polytope]>=0.1.17",
  "anemoi-utils>=0.4.36",
  "apscheduler",
  "cloudpickle",
  "earthkit-workflows>=0.5.1",
  "earthkit-workflows-anemoi>=0.3.12",
  "earthkit-workflows-pproc",
  "fastapi",
  "fastapi-users[oauth]",
  "fastapi-users-db-sqlalchemy",
  "httpx<1",
  "httpx-oauth",
  "jinja2",
  "mir-python==1.27.10.5",                            # NOTE once the pinning is fixed for mir-python, we can remove both this and multiolib pin
  "multiolib==2.7.1.5",
  "orjson",
  "pproc",
  "psutil",
  "pydantic-settings",
  "pyrsistent",
  "python-multipart",
  "qubed==0.1.12",
  "sse-starlette",
  "toml",
  "uvicorn",
]

optional-dependencies.all = [
  "forecast-in-a-box[plots,thermo,dev]",
]
optional-dependencies.plots = [
  "earthkit-plots>=0.3.5",
  "earthkit-plots-default-styles>=0.1",
]
optional-dependencies.thermo = [
  "thermofeel>=2.1.1",
]
optional-dependencies.webmars = [
  "ecmwf-api-client>=1.6.5",
]
urls."Source code" = "https://github.com/ecmwf/forecast-in-a-box"

[dependency-groups]
dev = [
  "pytest-asyncio",
  "pytest-xdist>=3.8",
  "ty==0.0.2",
  "prek",
]

[tool.setuptools_scm]
version_file = "forecastbox/_version.py"
write_to_template = '''# Do not change! Do not track in version control!
__version__ = "{version}"
'''
local_scheme = "no-local-version"
root = ".."
fallback_version = "0.0.0"
tag_regex = '^[vd](?P<version>\d+(?:\.\d+){0,2}[^\+]*)(?:\+.*)?$'

[tool.ruff]
line-length = 140

lint.ignore = [ "E731" ]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
testpaths = [ "tests" ]
addopts = "-n0 -s --log-disable findlibs"      # NOTE stick to 0 *or* have random server ports in the integration tests
# TODO once the diagnostic_/prognostic_ warning addressed, remove
filterwarnings = [
    "error",
    'ignore:diagnostic_variables is unsupported:deprecation.UnsupportedWarning',
    'ignore:prognostic_variables is unsupported:deprecation.UnsupportedWarning'
]

[tool.mypy]
plugins = "pydantic.mypy"

# TODO remove once released

[tool.uv.sources]
pproc = { git = "https://github.com/ecmwf/pproc" }
earthkit-workflows-pproc = { git = "https://github.com/ecmwf/earthkit-workflows-pproc" }

[tool.ty.rules]
possibly-missing-attribute = "ignore" # TODO actually fix all those occurrences instead
possibly-missing-import = "ignore"    # NOTE due to fastapi_users etc
